Update company and equipment entities, and adjust index and other files
The primary changes are in the company and equipment entities. All columns in the company and equipment entities have been given explicit names. In the entities, 'CreateDateColumn', 'UpdateDateColumn', and 'DeleteDateColumn' have been renamed to 'fecha_creacion', 'fecha_modificacion', and 'fecha_eliminacion' respectively. In addition, a new response-equipment-mini.mapper was created and methods for history and light were added in the equipment controller.