Add method to find company by ID
Added a new function that allows users to find a specific company by their ID in the CompanyService module. This function verifies the existence of the company in the database and raises an exception if the ID is not found. Corresponding changes are also made in the CompanyController to handle this new functionality.